Remove built bundle from repo and document build
