+++ apps/legal_discovery/src/components/PresentationSection.jsx
@@ -1,0 +1,20 @@
+import React, { useState } from "react";
+import { fetchJSON } from "../utils";
+function PresentationSection() {
+  const [path,setPath] = useState('');
+  const [slides,setSlides] = useState('');
+  const create = () => {
+    const slidesArr = slides.split('\n').map(l=>{const [t,...c]=l.split('|');return {title:t||'',content:c.join('|')||''};});
+    fetchJSON('/api/presentation',{method:'POST',headers:{'Content-Type':'application/json'},body:JSON.stringify({filepath:path,slides:slidesArr})}).then(alertResponse);
+  };
+  return (
+    <section className="card">
+      <h2>Presentation</h2>
+      <input type="text" value={path} onChange={e=>setPath(e.target.value)} className="w-full mb-2 p-2 rounded" placeholder="PPTX path" />
+      <textarea rows="4" value={slides} onChange={e=>setSlides(e.target.value)} className="w-full mb-2 p-2 rounded" placeholder="title|content per line" />
+      <button className="button-secondary" onClick={create}><i className="fa fa-slideshare mr-1"></i>Create/Update</button>
+    </section>
+  );
+}
+
+export default PresentationSection;
